--- a/test_plain_zarr_clean.yaml
+++ b/test_plain_zarr_clean.yaml
@@ -1,0 +1,25 @@
+general:
+  name: "Plain Zarr Test (2,121 MB/s)"
+
+input_data:
+  satellite:
+    optimal_time_steps: 6
+    zarr_path: "gs://gsoc-dakshbir/2024-02_nonhrv.zarr"
+    channels: ['IR_016', 'IR_039', 'IR_087', 'IR_097', 'IR_108', 'IR_120', 'IR_134', 'VIS006', 'VIS008', 'WV_062', 'WV_073']
+    time_resolution_minutes: 15
+    interval_start_minutes: -60
+    interval_end_minutes: 0
+    image_size_pixels_height: 128
+    image_size_pixels_width: 128
+    normalisation_constants:
+      IR_016: {mean: 0.5, std: 0.2}
+      IR_039: {mean: 0.5, std: 0.2}
+      IR_087: {mean: 0.5, std: 0.2}
+      IR_097: {mean: 0.5, std: 0.2}
+      IR_108: {mean: 0.5, std: 0.2}
+      IR_120: {mean: 0.5, std: 0.2}
+      IR_134: {mean: 0.5, std: 0.2}
+      VIS006: {mean: 0.5, std: 0.2}
+      VIS008: {mean: 0.5, std: 0.2}
+      WV_062: {mean: 0.5, std: 0.2}
+      WV_073: {mean: 0.5, std: 0.2}