## PR Description:

This Pull Request introduces a major new capability to ocf-data-sampler: high-performance satellite data loading directly from cloud storage using the icechunk library and optimized GCS streaming.

This work addresses the bottleneck of downloading large Zarr datasets by enabling a "cloud-native" approach. It provides a robust, version-controlled, and highly paerformant alternative to the existing plain Zarr loader, with benchmarks demonstrating throughput on par with or exceeding 14,000 MB/s.

## Key Features & Changes

### Ice Chunk Integration
Adds two new data loaders for satellite data:
- An optimized loader for standard cloud Zarr files
- A "true Ice Chunk" loader that leverages the icechunk library for versioning and transactional reads

### Configuration Model Update
The Satellite configuration model has been extended to support Ice Chunk-specific parameters (`icechunk_path`, `use_true_icechunk`, `icechunk_commit`, etc.).

### Data Conversion Pipeline
A new script is provided to convert existing OCF Zarr datasets into a high-performance Ice Chunk format with a compatible data structure.

### Comprehensive Benchmarking Suite
New scripts and utilities have been added to rigorously benchmark and compare the performance of different data loading strategies.

## File-by-File Breakdown

Here is a description of each new or modified file included in this PR:

### 1. Core Library Changes

#### [`[ocf_data_sampler/config/model.py]`](ocf_data_sampler/config/model.py)
- **Purpose:** Extends the Pydantic configuration model to support Ice Chunk
- **Details:** Adds fields like `icechunk_path`, `use_true_icechunk`, and optimization parameters to the Satellite model, allowing users to configure the new data loaders through YAML files

#### [`[ocf_data_sampler/load/icechunk_optimized.py]`](ocf_data_sampler/load/icechunk_optimized.py)
- **Purpose:** A high-performance loader for standard cloud-hosted Zarr files
- **Details:** Implements best practices for GCS streaming (e.g., gcsfs tuning, Dask configuration) to maximize throughput for Zarr datasets that are not managed by the icechunk library

#### [`[ocf_data_sampler/load/icechunk_optimized_with_ics.py]`](ocf_data_sampler/load/icechunk_optimized_with_ics.py)
- **Purpose:** The primary loader for "true" Ice Chunk repositories
- **Details:** This loader uses the icechunk library to interact with version-controlled datasets, allowing for transactional reads from specific branches or commits. It also handles the conversion of different Zarr data layouts into the format expected by the sampler

#### [`[ocf_data_sampler/load/load_dataset.py]`](ocf_data_sampler/load/load_dataset.py)
- **Purpose:** The central dispatcher for data loading
- **Details:** The logic in `get_dataset_dict` has been updated to act as a router. Based on the provided configuration (`zarr_path` vs. `icechunk_path`), it intelligently selects the appropriate loader (traditional, optimized cloud, or true Ice Chunk) for satellite data

### 2. Benchmarking & Utility Scripts

#### [`[ocf_data_sampler/torch_datasets/utils/benchmark.py]`](ocf_data_sampler/torch_datasets/utils/benchmark.py)
- **Purpose:** A new utility module to run and aggregate performance benchmarks
- **Details:** Provides the `OCFDataSamplerBenchmark` class, which programmatically runs loading tests, calculates average throughput, and provides a summary assessment

#### [`[scripts/benchmark_cli.py]`](scripts/benchmark_cli.py) 
- **Purpose:** A command-line interface for the benchmarking utility
- **Details:** Allows users to easily run a performance benchmark against any data source by simply providing a configuration file

#### [`[scripts/full_dataset_icechunk_conversion.py]`](scripts/full_dataset_icechunk_conversion.py)
- **Purpose:** A crucial tool for preparing data for the new pipeline
- **Details:** This script converts an existing OCF Zarr dataset (with its custom ocf_blosc2 codec) into a high-performance Ice Chunk repository. It correctly restructures the data with separate variables per channel and removes incompatible codecs, which is essential for achieving maximum read performance

#### [`[scripts/production_benchmark_comparison.py]`](scripts/production_benchmark_comparison.py)
- **Purpose:** A high-level script for direct performance comparison
- **Details:** This script runs an "apples-to-apples" benchmark comparing the existing plain Zarr loader against the new true Ice Chunk loader, using their respective optimal configurations. It provides a clear, final recommendation based on the results

### 3. Example Configuration Files

#### [`[production_icechunk_2024-02_config.yaml]`](production_icechunk_2024-02_config.yaml)
- **Purpose:** An example of a production-ready configuration for the Ice Chunk loader
- **Details:** Demonstrates how to configure the sampler to use a specific Ice Chunk repository, commit ID, and all 11 satellite channels

#### [`[test_plain_zarr_clean.yaml]`](test_plain_zarr_clean.yaml)
- **Purpose:** The baseline configuration for the existing plain Zarr loader
- **Details:** Used in comparison benchmarks to measure the performance of the current production data access method

## Final Benchmark Results

After fixing the data format to enable a true apples-to-apples comparison, the new Ice Chunk loader demonstrates exceptional performance, on par with and sometimes exceeding the existing plain Zarr implementation.

```
Production Comparison Results
========================================
Performance Comparison:
  Plain Zarr: 9261.93 MB/s
  Ice Chunk: 14256.93 MB/s
  Ice Chunk is 1.54x FASTER

Recommendation: Ice Chunk recommended for performance
```
This confirms that the new data access pattern is a viable and high-performance solution ready for production use.

## How to Test This PR
To verify the new functionality and reproduce the benchmark results, please follow these steps.

### Prerequisites:

Ensure all new dependencies are installed (e.g., icechunk).
Google Cloud authentication must be configured locally to access the GCS bucket (gsoc-dakshbir).
The source dataset gs://gsoc-dakshbir/2024-02_nonhrv.zarr must be available for the comparison benchmark.

**Step 1: Run the Main Production Benchmark**
This is the primary test. It runs a direct "apples-to-apples" comparison between the existing plain Zarr loader and the new, optimized Ice Chunk loader using the provided configuration files.

python scripts/production_benchmark_comparison.py

- **What it does:** The script warms up the GCS connection and then runs multiple timed samples for both the test_plain_zarr_clean.yaml and production_icechunk_2024-02_config.yaml configurations.

- **Expected Outcome:** The final output will be a summary table directly comparing the throughput of both methods. The results should be similar to those claimed in the PR description, showing a significant performance increase with the Ice Chunk loader.

### Production Comparison Results
========================================
Performance Comparison:
  Plain Zarr: ~9,200 MB/s
  Ice Chunk: ~14,000 MB/s
  Ice Chunk is ~1.5x FASTER

Recommendation: Ice Chunk recommended for performance

(Note: Absolute throughput may vary slightly based on network conditions, but the relative performance gain should be consistent.)


**Step 2: (Optional) Run an Individual Benchmark with the CLI Tool**
This demonstrates the flexibility of the new benchmark_cli.py utility for testing any configuration.

python scripts/benchmark_cli.py --config production_icechunk_2024-02_config.yaml --samples 3
python scripts/benchmark_cli.py --config test_plain_zarr_clean.yaml --samples 3


- **What it does**: Runs 3 benchmark samples using only the specified configuration.
- **Expected Outcome:** The script will print the average throughput for the icechunk_optimized method, which should be in the multi-GB/s range, confirming the high performance of the new loader independently.


**Step 3: (Optional) Verify the Data Conversion Pipeline**
This step validates the script that converts a standard OCF Zarr dataset into the high-performance Ice Chunk format. The PR already includes a pre-converted dataset specified in production_icechunk_2024-02_config.yaml, so this is for verification purposes.

python scripts/full_dataset_icechunk_conversion.py

- **What it does:** This script reads the 2024-02_nonhrv.zarr dataset, restructures it to have one variable per channel, removes the incompatible ocf_blosc2 codec, and writes it to a new version-controlled Ice Chunk repository in GCS.
- **Expected Outcome**: The script will print progress as it processes batches. Upon completion, it will output the GCS path of the newly created repository, a unique commit ID, and generate a new production_icechunk_..._config.yaml file in your local directory.