@dfulu 2025-12-12
This was very useful to explore, but since this is out of date with the current library and has gone stale I'm going to close it for now